Move adaptive LZMA decompression to a shared module
The parameter search lived in the v3 compressor, but nothing about it is
specific to v3. Move it to `backhand/src/lzma.rs` so other versions can use
it. No change in behavior.
