+++ backhand/src/lib.rs
@@ -58,6 +58,8 @@ type _ReadmeTest = ();
 
 pub mod error;
 mod kinds;
+#[cfg(feature = "v3_lzma")]
+mod lzma;
 pub mod traits;
 #[cfg(feature = "v3")]
 pub mod v3;
@@ -99,8 +101,8 @@ pub mod kind {
     pub use crate::kinds::{BE_V3_0, LE_V3_0};
     #[cfg(feature = "v3_lzma")]
     pub use crate::kinds::{
-        BE_V3_0_LZMA, BE_V3_1_LZMA_SWAP, LE_V3_0_LZMA, LE_V3_1_LZMA_SWAP, NETGEAR_BE_V3_0_LZMA,
-        NETGEAR_BE_V3_0_LZMA_STANDARD,
+        BE_V3_0_LZMA, BE_V3_1_LZMA_SWAP, LE_V3_0_LZMA, LE_V3_1_LZMA_SWAP,
+        NETGEAR_BE_V3_0_LZMA, NETGEAR_BE_V3_0_LZMA_STANDARD,
     };
 }
